Move the TriStateBoolean to the reused horizontal package - It's not bound to any vertical
